+++ CHANGELOG.md
@@ -29,6 +29,7 @@
     * applying filter before line processing if possible (spares a lot of time, ex: skip lines before WHOIS processing)
     * fix Python3.6: disabled logging messages from daemon (threw errors)
     * fix: processing velocity info (lines / s) showed garbage since threads implemented    
+    * fix: aggregate via Alt+a from main menu + any after-processing dialog (like resolve unknowns) caused the terminal to freeze (cannot use input() while another prompt toolkit session is active)
 
 ## 1.3.1 (2020-01-31)
 * thread processing